+++ packages/atlas/src/views/viewer/AuctionView/index.ts
@@ -1,0 +1,1 @@
+export * from './AuctionView'
