--- packages/atlas/src/views/viewer/PlaceBid/index.ts
@@ -1,1 +1,0 @@
-export * from './PlaceBid'
